Calculate totals using annotate, not aggregate
For reports like "all the accounts ever", this can
increase performance by a factor of about three
and reduce the number of database queries by a
factor of about a hundred (aka the number of rows
in the table produced).

It is substantially more complicated, though.
The original code is in a "slow = False; if slow:"
block. In theory, it'd be good to keep both codepaths
working and test them against each other.